+++ R/importRaster.R
@@ -1,34 +1,33 @@
 #' Import generic raster file
 #'
-#' The function import the first layer of a generic raster file. Data are
-#' imported as an array of x, y, z columns.
+#' A function to import the first layer of a generic raster file.
 #'
-#' Supported files include those managed by the \pkg{raster} package (as
-#' netcdf),
-#'
-#' Destaggering is useful for importing data from the SPRAY model and it is not
-#' applied by default.
-#'
-#' An optional summary output can be printed by setting the `verbose` parameter.
-#'
-#'
-#' @param file The raster file to be imported.
-#' @param k A numerical factor to be applied to x and y coordinates (default =
-#'   1).
-#' @param kz A numerical factor to be applied to z values (default = 1).
-#' @param dx Shifts x coordinates by dx (default = 0).
-#' @param dy float. Shift y coordinates by dy (default = 0).
+#' @param file character. Path to the raster file.
+#' @param k numeric. Factor applied to x and y coordinates (default = 1).
+#' For example, it can be used to convert the grid coordinates from km
+#' to m (k = 1000).
+#' @param kz numeric. Factor applied to the variable values (default = 1).
+#' @param dx numeric. Constant to shift x coordinates (default = 0).
+#' @param dy numeric. Constant to shift y coordinates (default = 0).
 #' @param destaggering Use `TRUE` to apply destaggering to X and Y coordinates
-#'   (default = FALSE).
-#' @param variable The name of the variable to be imported.
-#' @param verbose If `TRUE`, prints out basic statistics (default = FALSE).
+#'   (default = FALSE). See the `Details` section.
+#' @param variable character. The name of the variable to be imported.
+#' @param verbose logical. If `TRUE`, prints out basic statistics (default = FALSE).
 #'
 #' @details
-#' This function is based on the `terra` package and it can import any format
-#' managed by it.
+#' This function is based on the \pkg{terra} package and it can import any format
+#' managed by it as NetCDF.
+#'
+#' Destaggering applies a shift equal to half grid size in both horizontal
+#' directions. It is useful for importing data from the SPRAY air quality dispoersion
+#' model and it is not applied by default.
+#'
+#' An optional summary output can be printed out by setting the `verbose` parameter
+#' to `TRUE`.
 #'
 #'
-#' @return It returns a dataframe with x, y and z columns.
+#' @return A data.frame with x, y and z columns for the grid cells coordiantes
+#' and the variable value.
 #'
 #' @seealso [importADSOBIN()], [importSurferGrd()]
 #'
@@ -39,9 +38,10 @@
 #' @examples
 #' \dontrun{
 #' # Import binary (netcdf) file and convert coordinates from km to m,
-#' # without destaggering:
+#' # without destaggering. Variable name is "NOx".
 #' mydata <- importRaster(
 #'     file = "/path_to_file/filename.nc",
+#'     variable = "NOx",
 #'     k = 1000,
 #'     destaggering = FALSE
 #' )
@@ -50,26 +50,29 @@
 #' # with shift of 100 m in both directions:
 #' mydata <- importRaster(
 #'     file = "/path_to_file/filename.nc",
+#'     variable = "pm10",
 #'     k = 1000,
 #'     dx = 100,
 #'     dy = 100
 #' )
 #' }
-importRaster <- function(file = file.choose(),
-                         k = 1,
-                         kz = 1,
-                         dx = 0,
-                         dy = 0,
-                         destaggering = FALSE,
-                         variable = NULL,
-                         verbose = FALSE) {
+importRaster <- function(
+    file = file.choose(),
+    k = 1,
+    kz = 1,
+    dx = 0,
+    dy = 0,
+    destaggering = FALSE,
+    variable = NULL,
+    verbose = FALSE
+) {
 
-    if (is.null(variable)) {
+    if (missing(variable)) {
         t <- terra::rast(file)
-        variable <- as.character(t@data@names)
-    } else {
-        t <- terra::rast(file, subds = as.character(variable))
+        variables <- as.character(names(t))
+        stop("Missing variables. Choose one from: ", list(variables))
     }
+    t <- terra::rast(file, subds = as.character(variable))
 
     # Apply conversion factor
     terra::xmax(t) <- terra::xmax(t) * k
@@ -94,20 +97,14 @@ importRaster <- function(file = file.choose(),
 
     # Print some values
     if (verbose == TRUE) {
-        cat("\nRaster statistics -----------------------------------------------")
         xvalues <- c(terra::xmin(t), terra::xmax(t), terra::res(t)[1])
-        cat(sprintf("\n%8s (min, max, dx)  :", "X"))
-        cat(sprintf(fmt = "%12.3f", xvalues))
-
         yvalues <- c(terra::ymin(t), terra::ymax(t), terra::res(t)[2])
-        cat(sprintf("\n%8s (min, max, dy)  :", "Y"))
-        cat(sprintf(fmt = "%12.3f", yvalues))
-
         zvalues <- c(terra::global(t, min), terra::global(t, max), terra::global(t, mean))
-        cat(sprintf("\n%8s (min, max, mean):", variable))
-        cat(sprintf(fmt = "%12.2e", zvalues))
-
-        cat("\n-----------------------------------------------------------------\n")
+        message("Raster statistics -----------------------------------------------")
+        message(sprintf("%8s (min, max, dx)  : %12.3f %12.3f %12.3f", "X", xvalues[1], xvalues[2], xvalues[3]))
+        message(sprintf("%8s (min, max, dy)  : %12.3f %12.3f %12.3f", "Y", yvalues[1], yvalues[2], yvalues[3]))
+        message(sprintf("%8s (min, max, mean): %12.2e %12.2e %12.2e", variable, zvalues[1], zvalues[2], zvalues[3]))
+        message("-----------------------------------------------------------------")
     }
 
     # Export dataframe with x, y, x columns