+++ tests/testthat/test_contourPlot2.R
@@ -1,7 +1,14 @@
 test_that("contourPlot2 is a ggplot2 object", {
     data("volcano")
-    volcano <- reshape2::melt(volcano)
-    names(volcano) <- c("x", "y", "z")
-    v <- contourPlot2(volcano)
+    volcano <- as.data.frame(volcano)
+    volcano3d <- reshape(
+        volcano,
+        direction = "long",
+        varying = list(1:61),
+        idvar = "x",
+        timevar = "y",
+        v.names = "z"
+    )
+    v <- contourPlot2(volcano3d)
     expect_s3_class(v, "ggplot")
 })