+++ NEWS.md
@@ -19,7 +19,7 @@
   English.
 - `plotAvgRad()` now includes options to change the y-axis label and the
   title.
-- `plotAvgTemp()` and `plotStabilityClass()` gain `date` argument to specify the
+- `plgtAvgTemp()` and `plotStabilityClass()` gain `date` argument to specify the
   name of the column with *datetime* data.
 
 ## Minor improvements and fixes