+++ NEWS.md
@@ -1,3 +1,5 @@
+# simulariatools (development version)
+
 # simulariatools 3.0.0
 
 ## Breaking changes